Extract join-key resolution into a JoinKeyResolver collaborator
The ReadModelScenario join backfill logic (root-join resolution, child-join
re-anchoring, phantom avoidance, child-path/indexer construction) had
accreted onto ProjectionReadModelProcessor, pushing it past the file-size
guideline and turning ProcessSingleEvent into a method that needed section
comments to be read.

Move that cohesive concern into a dedicated JoinKeyResolver so
ProcessSingleEvent reads as orchestration and the join "how" lives in one
place. Pure refactor — the method bodies are moved verbatim and the
sink-before-threaded-state ordering is unchanged; the full harness spec
suite is identical before and after.
